+++ src/network/packets/handlers/ModalFormResponse.js
@@ -10,7 +10,7 @@
  * Copyright 2023 andriycraft
  * Github: https://github.com/andriycraft/GreenFrogMCBE
  */
-const Events = require("../../../server/Events");
+const Events = require("../../../plugin/Events");
 
 class ModalFormResponse extends require("./Handler") {
   handle(server, client, packet) {
